refactor(pools): move util functions to test-util file
